fix(events): events promotion structure update; multiple sidebar
