+++ src/components/Button/ButtonStyle.ts
@@ -1,0 +1,21 @@
+import React from "react";
+
+export const normalButtonStyle: React.CSSProperties = {
+  background: "white",
+  color: "black",
+  border: "1px solid #233F55",
+  padding: "6px 12px 6px 12px",
+  fontSize: "12px",
+  lineHeight: "18px",
+  fontWeight: 700,
+};
+
+export const activatedButtonStyle: React.CSSProperties = {
+  background: "#233F55",
+  color: "white",
+  border: "1px solid white",
+  padding: "6px 12px 6px 12px",
+  fontSize: "12px",
+  lineHeight: "18px",
+  fontWeight: 700,
+};
