+++ src/components/Sidebar/Sidebar.tsx
@@ -54,3 +54,61 @@ export const Sidebar: React.FC<SidebarProps> = ({
     </div>
   );
 };
+
+type MultipleSidebarsProps = {
+  sidebars: Array<{
+    title: string;
+    sidebarItems: SubItems[];
+  }>;
+  activatedItems: string;
+  setActivatedItems: (activatedItems: string) => void;
+};
+export const MultipleSidebars: React.FC<MultipleSidebarsProps> = (
+  props: MultipleSidebarsProps
+) => {
+  const { sidebars, activatedItems, setActivatedItems } = props;
+  return (
+    <div className="border-2 border-inherit p-[24px]">
+      {sidebars.map((sidebar) => {
+        const { title, sidebarItems } = sidebar;
+        return (
+          <div className="flex flex-col gap-[24px] mb-[24px]">
+            <p className="text-heading-l">{title}</p>
+            <div className="w-full flex flex-col gap-[24px]">
+              {sidebarItems.map((item, i) => {
+                const { subTitle, imgUrl, enum: navItemEnum } = item;
+                const isActivated = activatedItems === navItemEnum;
+                return (
+                  <div
+                    key={i}
+                    className={`transition-all duration-300 ease-in-out ${
+                      isActivated
+                        ? "bg-[#F2F2EF] border-[16px] border-[#F2F2EF]"
+                        : "bg-transparent border-[0px] border-transparent"
+                    }`}
+                    onClick={() => {
+                      setActivatedItems(navItemEnum);
+                    }}
+                  >
+                    <div className="flex flex-row h-[106px] gap-[24px] w-full">
+                      <img
+                        className="w-auto h-[106px] object-contain"
+                        src={
+                          process.env.PUBLIC_URL + "/assets/images/" + imgUrl
+                        }
+                        alt={imgUrl}
+                      />
+                      <div className="text-highlight-m text-black py-[8px] pr-[16px]">
+                        {subTitle}
+                      </div>
+                    </div>
+                  </div>
+                );
+              })}
+            </div>
+          </div>
+        );
+      })}
+    </div>
+  );
+};
