+++ src/components/index.ts
@@ -13,3 +13,4 @@ export * from "./InternalBackButton";
 export * from "./Table";
 export * from "./Link";
 export * from "./FileTemplate";
+export * from "./Button";
