--- src/pages/EventsLanding/index.ts
@@ -1,2 +1,0 @@
-export * from "./EventsLanding";
-export * from "./AwardScheme";
