+++ src/pages/EventsPromotion/EventsLanding/AwardScheme/index.ts
@@ -1,0 +1,2 @@
+export * from "./AwardScheme";
+export * from "./AwardScheme23to24";
