+++ src/pages/EventsPromotion/EventsLanding/index.ts
