+++ src/pages/EventsPromotion/Publication/Newspaper/index.ts
@@ -1,0 +1,1 @@
+export * from "./Newsletter";
