+++ src/pages/EventsPromotion/Publication/index.ts
@@ -1,0 +1,1 @@
+export * from "./Newspaper";
