+++ src/pages/index.ts
@@ -3,7 +3,6 @@ export * from "./GeneralPublic";
 export * from "./EducatorsStudents";
 export * from "./Industry";
 export * from "./ServiceUsers";
-export * from "./EventsLanding";
 export * from "./SeminarsWorkshops";
 export * from "./StudentCompetitions";
 export * from "./ResourcesReports";
@@ -12,3 +11,4 @@ export * from "./AboutUs";
 export * from "./TCSector";
 export * from "./WhatsNew";
 export * from "./AboutSite";
+export * from "./EventsPromotion";
